schema: refactor data models, clean up schema & remove old migrations
- Convert cals_per_100g & weight fields to floats
- Update repos/forms/etc to reflect this
- Remove nonsensical min/max constants for promotion_threshold (set
  internally)
- Add Pillars model and relationship to start drafting that

+++ app/modules/groceries/repository.py
@@ -37,7 +37,7 @@ class ProductRepository(BaseRepository[Product]):
         net_weight: Decimal,
         unit_type: UnitEnum,
         barcode: str | None,
-        calories_per_100g: Decimal | None,
+        calories_per_100g: float | None,
     ) -> Product:
         product = Product(
             user_id=self.user_id,
